feat: 🎸 separate the hamburger component
