+++ .eslintrc.json
@@ -13,6 +13,7 @@
     "consistent-return": "off",
     "no-underscore-dangle": "off",
     "arrow-body-style": "off",
+    "react/require-default-props": "off",
     "react/function-component-definition": [
       2,
       { "namedComponents": "arrow-function" }
