+++ src/components/Header/index.tsx
@@ -5,6 +5,7 @@ import HamburgerBotton from '@components/Header/HamburgerBotton';
 import LogoButton from '@components/Header/LogoButton';
 import NotificationButton from '@components/Header/NotificationButton';
 import Sidebar from '@components/Sidebar';
+import useWindowSize from '@hooks/use-window-size';
 import { useAuthContext } from 'src/store/AuthProvider/AuthProvider';
 
 export type UserProfileButtonProps = {
@@ -13,6 +14,7 @@ export type UserProfileButtonProps = {
 
 const Header = () => {
   const { isLoggedin, logout } = useAuthContext();
+  const { width } = useWindowSize();
   const handleSignOut: UserProfileButtonProps['onSignOut'] = () => {
     logout();
   };
@@ -22,19 +24,18 @@ const Header = () => {
       <div className="flex h-28 w-full max-w-[81rem] items-center justify-between bg-surface xl:h-36">
         <LogoButton />
         <div className="flex items-center justify-center">
-          <div className="hidden xl:block">
-            <EmailButton />
-            <NotificationButton />
-          </div>
-          <div className="hidden xl:block">
-            <HamburgerBotton
-              isLoggedin={isLoggedin}
-              onSignOut={handleSignOut}
-            />
-          </div>
-          <div className="xl:hidden">
+          {width > 1366 ? (
+            <>
+              <EmailButton />
+              <NotificationButton />
+              <HamburgerBotton
+                isLoggedin={isLoggedin}
+                onSignOut={handleSignOut}
+              />
+            </>
+          ) : (
             <Sidebar isLoggedin={isLoggedin} onSignOut={handleSignOut} />
-          </div>
+          )}
         </div>
       </div>
     </header>
