+++ src/components/Sidebar/index.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import Hamburger from '@components/Hamburger';
 import {
   SITEMAP,
   registerAndLoginList,
@@ -12,12 +13,8 @@ import {
   DialogTrigger,
 } from '@components/ui/dialog';
 import { Separator } from '@components/ui/separator';
-import { SVGMotionProps, motion, useCycle } from 'framer-motion';
-import Image from 'next/image';
+import { motion, useCycle } from 'framer-motion';
 import Link from 'next/link';
-
-const defaultAvatar = '/header__defaultAvatar.svg';
-const fakeAvatar = '/header__fakeAvatar.svg';
 
 type SidebarProps = {
   isLoggedin: boolean;
@@ -29,16 +26,6 @@ type List = {
   icon?: JSX.Element;
   url: string;
 };
-
-const Path = (props: SVGMotionProps<SVGPathElement>) => (
-  <motion.path
-    fill="transparent"
-    strokeWidth="2"
-    stroke="#403E3D"
-    strokeLinecap="round"
-    {...props}
-  />
-);
 
 const menuAnimation = {
   open: {
@@ -72,51 +59,14 @@ const Sidebar = ({ isLoggedin, onSignOut }: SidebarProps) => {
   return (
     <Dialog open={isOpen} onOpenChange={changeIsOpen}>
       <motion.nav initial={false} animate={isOpen ? 'open' : 'closed'}>
-        <DialogTrigger
-          asChild
-          className={`
-          ${isOpen && 'fixed right-[0.75rem] top-[1.25rem] z-[51]'} 
-          mx-[0.1875rem] flex items-center justify-center rounded-full border border-primary bg-surface p-3 data-[state=open]:mx-0 data-[state=open]:border-4`}
-        >
-          <div>
-            <div className="p-2">
-              <svg width="24" height="24" viewBox="0 0 24 24">
-                <Path
-                  variants={{
-                    closed: { d: 'M 3 6 L 21 6' },
-                    open: { d: 'M 3 18 L 21 6' },
-                  }}
-                  initial={false}
-                  animate={isOpen ? 'open' : 'closed'}
-                />
-                <Path
-                  d="M 3 12 L 21 12"
-                  variants={{
-                    closed: { opacity: 1 },
-                    open: { opacity: 0 },
-                  }}
-                  transition={{ duration: 0.1 }}
-                  initial={false}
-                  animate={isOpen ? 'open' : 'closed'}
-                />
-                <Path
-                  variants={{
-                    closed: { d: 'M 3 18 L 21 18' },
-                    open: { d: 'M 3 6 L 21 18' },
-                  }}
-                  initial={false}
-                  animate={isOpen ? 'open' : 'closed'}
-                />
-              </svg>
-            </div>
-            <Image
-              className="ml-2"
-              src={isLoggedin ? fakeAvatar : defaultAvatar}
-              alt="user"
-              width={40}
-              height={40}
-            />
-          </div>
+        <DialogTrigger>
+          <Hamburger
+            className={
+              isOpen ? 'fixed right-[0.75rem] top-[1.25rem] z-[51]' : ''
+            }
+            isOpen={isOpen}
+            isLoggedin={isLoggedin}
+          />
         </DialogTrigger>
         <DialogContent
           hideCloseButton
